#100: Fixes - Fixed a bug where half rating would crash the app now parsed to int becoming 0 - Fixed a bug where scrolling down would always refresh the page added onScrollListener - Code tidy

+++ MobileApp/app/src/main/java/pt/ipleiria/estg/dei/waypinpoint/ReviewDetailsActivity.java
@@ -57,7 +57,7 @@ public class ReviewDetailsActivity extends AppCompatActivity implements ReviewLi
         fabSave.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                float rating = rbScore.getRating();
+                int rating = (int) rbScore.getRating();
                 String message = etMessage.getText().toString().trim();
 
                 if (rating == 0) {
